miner_detector: BUGFIX - gcc10 fails because of narrowing
This patch fixes the error message:
narrowing conversion of ‘3735928495’ from ‘unsigned int’ to ‘int’

+++ miner_detector/miner_detector.h
@@ -91,7 +91,7 @@
 /**
  * Used when not using active test.
  */
-#define STRATUM_NOT_USED 0xdeadbeaf
+#define STRATUM_NOT_USED ((int) (0xdeadbeaf))
 
 /**
  * Used to identify permanent records in blacklist/whitelist DB.
